+++ src/ServiceControl.Transports.SqlServer/NameHelper.cs
@@ -14,5 +14,15 @@
 
             return prefix + name.Replace(suffix, suffix + suffix) + suffix;
         }
+
+        public static string Unquote(string quotedString)
+        {
+            if (!quotedString.StartsWith(prefix) || !quotedString.EndsWith(suffix))
+            {
+                return quotedString;
+            }
+            return quotedString
+                .Substring(prefix.Length, quotedString.Length - prefix.Length - suffix.Length).Replace(suffix + suffix, suffix);
+        }
     }
 }
